# pycxids/core/http_binding/http_binding_openapi.yaml
openapi: "3.0.0"
info:
  description: "IDS spec"
  version: "0.8"
  title: "IDS http binding"
  
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"


tags:
- name: "Catalog"
  description: "Catalog endpoints"

- name: "Negotiation"
  description: "Negotiation Process - Provider side"

- name: "Negotiation Callbacks"
  description: "Negotiation Process Callbacks - Consumer side"

- name: "Transfer"
  description: "Transfer Process"

paths:
  /catalog/request:
    post:
      tags:
      - "Catalog"
      requestBody:
        required: True
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CatalogRequestMessage"
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DcatCatalog"
          description: "Result"
  /catalog/datasets/{id}:
    get:
      tags:
        - "Catalog"
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: "Result"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DcatDataset"

  /negotiations/{id}:
    get:
      tags:
        - "Negotiation"
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: "Result"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContractNegotiation"
  
  /negotiations/request:
    post:
      tags:
        - "Negotiation"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContractRequestMessage"
      responses:
        "201":
          description: "The provider connector must return an HTTP 201 (Created) response with the location header set to the location of the contract negotiation and a body containing the ContractNegotiation message"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContractNegotiation"
  /negotiations/{id}/request:
    post:
      tags:
        - "Negotiation"
      description: "The consumer must include the processId. The consumer must include either the offer or offerId property."
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContractRequestMessage"
      responses:
        "200":
          description: "If the message is successfully processed, the provider connector must return and HTTP 200 (OK) response. The response body is not specified and clients are not required to process it."
  /negotiations/{id}/events:
    post:
      tags:
        - "Negotiation"
      description: "The consumer must include the processId. The consumer must include either the offer or offerId property."
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContracNegotiationEventMessage"
      responses:
        "200":
          description: "Result"
  /negotiations/{id}/agreement/verification:
    post:
      tags:
        - "Negotiation"
      description: "TODO"
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContractAgreementVerificationMessage"
      responses:
        "200":
          description: "Result"
  /negotiations/{id}/termination:
    post:
      tags:
        - "Negotiation"
      description: "TODO"
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContractNegotiationTerminationMessage"
      responses:
        "200":
          description: "Result"

  /negotiations/{id}/offer:
    post:
      tags:
        - "Negotiation Callbacks"
      description: "Callback on the Consumer side"
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContractOfferMessage"
      responses:
        "200":
          description: "Result"
  /negotiations/{id}/agreement:
    post:
      tags:
        - "Negotiation Callbacks"
      description: "Callback on the Consumer side"
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContractAgreementMessage"
      responses:
        "200":
          description: "Result"

  /transfers/request:
    post:
      tags:
        - "Transfer"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransferRequestMessage"
      responses:
        "201":
          description: "The provider connector must return an HTTP 201 (Created) response with the location header set to the location of the transfer process and a body containing the TransferProcess message"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransferProcess"


components:
  schemas:
    DspaceFilter:
      type: object
    CatalogRequestMessage:
      type: object
      properties:
        "@context":
          type: string
          default: "https://w3id.org/dspace/v0.8/context.json"
        "@type":
          type: string
          default: "dspace:CatalogRequestMessage"
        filter:
          $ref: "#/components/schemas/DspaceFilter"

    ContractRequestMessage:
      type: object
      properties:
        "@context":
          type: string
          default: "https://w3id.org/dspace/v0.8/context.json"
        "@id":
          type: string
        dspace:dataset:
          type: string
          description: "@id of the dataset"
        dspace:processId:
          type: string
          description: "TODO: Deprecated? To be removed?"
        dspace:offer:
          $ref: "#/components/schemas/OdrlOffer"
        dspace:callbackAddress:
          type: string
    
    ContracNegotiationEventMessage:
      type: object
      properties:
        "@context":
          type: string
          default: "https://w3id.org/dspace/v0.8/context.json"
        "@type":
          type: string
          default: "dspace:ContractNegotiationEventMessage"
        "dspace:processId":
          type: string
        "dspace:eventType":
          type: string
          enum:
            - FINALIZED
            - ACCEPTED
            - TERMINATED

    ContractAgreementVerificationMessage:
      allOf:
        - $ref: "#/components/schemas/JsonLd"
        - type: object
          description: "TODO: not finished yet"
          properties:
            "@type":
              type: string
              default: "dspace:ContractAgreementVerificationMessage"
            "dspace:processId":
              type: string
    ContractNegotiationTerminationMessage:
      allOf:
        - $ref: "#/components/schemas/JsonLd"
        - type: object
          description: "TODO: not finished yet"
          properties:
            "@type":
              type: string
              default: "dspace:ContractNegotiationTerminationMessage"
            "dspace:processId":
              type: string
            "dspace:code":
              type: string
              description: "TODO: not documented?"
            "dspace:reason":
              description: "TODO: can be a link / @id too"
              type: array
              items:
                $ref: "#/components/schemas/LanguageValue"

    ContractOfferMessage:
      allOf:
        - $ref: "#/components/schemas/JsonLd"
        - type: object
          description: "Used for Tranistion Requested -> Offered from Provider -> Consumer Callback endpoint"
          properties:
            "@type":
              type: string
              default: "dspace:ContractOfferMessage"
            dspace:processId:
              type: string
            dspace:offer:
              $ref: "#/components/schemas/OdrlOffer"
            dspace:callbackAddress:
              type: string

    ContractAgreementMessage:
      allOf:
        - $ref: "#/components/schemas/JsonLd"
        - type: object
          description: "Used for Tranistion Requested -> Agreed from Provider -> Consumer Callback endpoint"
          properties:
            "@type":
              type: string
              default: "dspace:ContractAgreementMessage"
            dspace:processId:
              type: string
            dspace:agreement:
              $ref: "#/components/schemas/OdrlAgreement"
            dspace:callbackAddress:
              type: string

    
    ContractNegotiation:
      allOf:
        - $ref: "#/components/schemas/JsonLd"
        - type: object
          properties:
            "@type":
              type: string
              default: "dspace:ContractNegotiation"
            "dspace:processId":
              type: string
            "dspace:state":
              $ref: "#/components/schemas/TransferState"

    TransferRequestMessage:
      allOf:
        - $ref: "#/components/schemas/JsonLd"
        - type: object
          description: "Used for Transfer Tranistion Start -> Requested from Consumer -> Provider"
          properties:
            "@type":
              type: string
              default: "dspace:TransferRequestMessage"
            dspace:agreementId:
              type: string
              description: "The agreementId property refers to an existing contract agreement between the consumer and provider."
            dct:format:
              type: string
              description: "The dct:format property is a format specified by a Distribution for the Asset associated with the agreement. This is generally obtained from the provider Catalog."
            dspace:dataAddress:
              type: string
              description: "The dataAddress property must only be provided if the dct:format requires a push transfer."
            dspace:callbackAddress:
              type: string

    TransferStartMessage:
      allOf:
        - $ref: "#/components/schemas/JsonLd"
        - type: object
          description: "Used for Transfer Tranistion Requested -> Started from Provider -> Consumer"
          properties:
            "@type":
              type: string
              default: "dspace:TransferStartMessage"
            dspace:processId:
              type: string
            dspace:dataAddress:
              type: string
    TransferProcess:
      allOf:
        - $ref: "#/components/schemas/JsonLd"
        - type: object
          properties:
            "@type":
              type: string
              default: "dspace:TransferProcess"
            dspace:processId:
              type: string
            dspace:state:
              $ref: "#/components/schemas/TransferState"

    NegotiationState:
      type: string
      enum:
        - REQUESTED
        - OFFERED
        - ACCEPTED
        - AGREED
        - VERIFIED
        - FINALIZED
        - TERMINATED

    TransferState:
      type: string
      enum:
        - REQUESTED
        - STARTED
        - TERMINATED
        - COMPLETED
        - SUSPENDED
      

    LanguageValue:
      type: object
      properties:
        "@value":
          type: string
        "@language":
          type: string
    JsonLd:
      type: object
      properties:
        "@context":
          type: string
          default: "https://w3id.org/dspace/v0.8/context.json"
        "@id":
          type: string
      

    DcatDataset:
      type: object
      properties:
        "@id":
          type: string
        "@type":
          type: string
        dct:title:
          type: string
        dct:description:
          type: string
        dct:keyword:
          type: array
          items:
            type: string
        odrl:hasPolicy:
          type: array
          items:
            $ref: "#/components/schemas/OdrlPolicy"
        dcat:distribution:
          $ref: "#/components/schemas/DcatDistribution"
    DcatCatalog:
      type: object
      properties:
        "@context":
          type: string
        "@id":
          type: string
        "@type":
          type: string
        "dct:title":
          type: string
        "dct:publisher":
          type: string
        "dcat:keyword":
          type: array
          items:
            type: string
        dcat:service:
          $ref: "#/components/schemas/DcatService"
        dcat:dataset:
          type: array
          items:
            $ref: "#/components/schemas/DcatDataset"
    DcatDistribution:
      type: object
      properties:
        "@type":
          type: string
        dct:format:
          type: string
        dcat:accessService:
          type: string
    DcatService:
      type: object
      properties:
        "@id":
          type: string
        "@type":
          type: string
        dct:terms:
          type: string
        dct:endpointUrl:
          type: string


    OdrlOperand:
      type: object
      properties:
        value:
          type: string
    OdrlConstraint:
      type: object
      properties:
        leftOperand:
          $ref: "#/components/schemas/OdrlOperand"
        rightOperand:
          $ref: "#/components/schemas/OdrlOperand"
        operator:
          type: string
        
    OdrlRule:
      type: object
      properties:
        action:
          type: string
        constraint:
          $ref: "#/components/schemas/OdrlConstraint"
        duty:
          type: array
          items:
            type: string # TODO: what is this exactly?
    OdrlPolicy:
      description: "In IDS http binding explicitly does NOT have a target, because this is derived from the enclosing context!"
      type: object
      properties:
        "@id":
          type: string
        "@type":
          type: string
          default: "odrl:Policy"
        permission:
          type: array
          items:
            $ref: "#/components/schemas/OdrlRule"
        prohibition:
          type: array
          items:
            $ref: "#/components/schemas/OdrlRule"
        obligation:
          type: array
          items:
            $ref: "#/components/schemas/OdrlRule"
    OdrlOffer:
      description: "Only addition compared to the Policy: the target field"
      allOf:
        - $ref: "#/components/schemas/OdrlPolicy"
        - type: object
          properties:
            "@type":
              type: string
              default: "odrl:Offer"
            odrl:target:
              type: string
    OdrlAgreement:
      description: "Agreement"
      allOf:
        - $ref: "#/components/schemas/OdrlPolicy"
        - type: object
          properties:
            "@type":
              type: string
              default: "odrl:Agreement"
            odrl:target:
              type: string
            dspace:timestamp:
              $ref: "#/components/schemas/DspaceTimestamp"
            dspace:consumerId:
              type: string
            dspace:providerId:
              type: string
    DspaceTimestamp:
      description: "xsd:dateTime"
      properties:
        "@type":
          type: string
          default: "xsd:dateTime"
        "@value":
          type: string